+++ ui/pages/bridge/prepare/prepare-bridge-page.tsx
@@ -19,6 +19,7 @@ import {
   BRIDGE_QUOTE_MAX_RETURN_DIFFERENCE_PERCENTAGE,
   type GenericQuoteRequest,
   getNativeAssetForChainId,
+  isNativeAddress,
 } from '@metamask/bridge-controller';
 import type { BridgeToken } from '@metamask/bridge-controller';
 import {
@@ -48,6 +49,7 @@ import {
   isBridgeSolanaEnabled,
   getIsToOrFromSolana,
   getQuoteRefreshRate,
+  getHardwareWalletName,
 } from '../../../ducks/bridge/selectors';
 import {
   AvatarFavicon,
@@ -121,7 +123,7 @@ import { useTokenAlerts } from '../../../hooks/bridge/useTokenAlerts';
 import { useDestinationAccount } from '../hooks/useDestinationAccount';
 import { Toast, ToastContainer } from '../../../components/multichain';
 import { MultichainNetworks } from '../../../../shared/constants/multichain/networks';
-import { isSwapsDefaultTokenAddress } from '../../../../shared/modules/swaps.utils';
+import { useIsTxSubmittable } from '../../../hooks/bridge/useIsTxSubmittable';
 import { BridgeInputGroup } from './bridge-input-group';
 import { BridgeCTAButton } from './bridge-cta-button';
 import { DestinationAccountPicker } from './components/destination-account-picker';
@@ -187,6 +189,8 @@ const PrepareBridgePage = () => {
   const keyring = useSelector(getCurrentKeyring);
   // @ts-expect-error keyring type is wrong maybe?
   const isUsingHardwareWallet = isHardwareKeyring(keyring.type);
+  const hardwareWalletName = useSelector(getHardwareWalletName);
+  const isTxSubmittable = useIsTxSubmittable();
   const locale = useSelector(getIntlLocale);
 
   const ticker = useSelector(getNativeCurrency);
@@ -214,7 +218,13 @@ const PrepareBridgePage = () => {
   const {
     filteredTokenListGenerator: toTokenListGenerator,
     isLoading: isToTokensLoading,
-  } = useTokensWithFiltering(toChain?.chainId ?? fromChain?.chainId, fromToken);
+  } = useTokensWithFiltering(
+    toChain?.chainId ?? fromChain?.chainId,
+    fromToken,
+    selectedDestinationAccount !== null && 'id' in selectedDestinationAccount
+      ? selectedDestinationAccount.id
+      : undefined,
+  );
 
   const { flippedRequestProperties } = useRequestProperties();
   const trackCrossChainSwapsEvent = useCrossChainSwapsEventTracker();
@@ -274,6 +284,8 @@ const PrepareBridgePage = () => {
           setFromToken({
             ...srcAsset,
             chainId: srcChainId,
+            // TODO: Fix in https://github.com/MetaMask/metamask-extension/issues/31880
+            // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
             image: srcAsset.icon || srcAsset.iconUrl || '',
             address: srcAsset.address,
           }),
@@ -446,9 +458,8 @@ const PrepareBridgePage = () => {
   }, []);
 
   const occurrences = Number(toToken?.occurrences ?? 0);
-  const toTokenIsNotDefault =
-    toToken?.address &&
-    !isSwapsDefaultTokenAddress(toToken?.address, toChain?.chainId as string);
+  const toTokenIsNotNative =
+    toToken?.address && !isNativeAddress(toToken?.address);
 
   const isSolanaBridgeEnabled = useSelector(isBridgeSolanaEnabled);
 
@@ -527,6 +538,8 @@ const PrepareBridgePage = () => {
           amountFieldProps={{
             testId: 'from-amount',
             autoFocus: true,
+            // TODO: Fix in https://github.com/MetaMask/metamask-extension/issues/31880
+            // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
             value: fromAmount || undefined,
           }}
           isTokenListLoading={isFromTokensLoading}
@@ -660,9 +673,13 @@ const PrepareBridgePage = () => {
                   }
             }
             customTokenListGenerator={
+              // TODO: Fix in https://github.com/MetaMask/metamask-extension/issues/31880
+              // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
               toChain || isSwap ? toTokenListGenerator : undefined
             }
             amountInFiat={
+              // TODO: Fix in https://github.com/MetaMask/metamask-extension/issues/31880
+              // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
               activeQuote?.toTokenAmount?.valueInCurrency || undefined
             }
             amountFieldProps={{
@@ -805,6 +822,34 @@ const PrepareBridgePage = () => {
               </Footer>
             </Column>
           </Row>
+          {isUsingHardwareWallet &&
+            isTxSubmittable &&
+            hardwareWalletName &&
+            activeQuote && (
+              <BannerAlert
+                marginInline={4}
+                marginBottom={3}
+                title={t('hardwareWalletSubmissionWarningTitle')}
+                textAlign={TextAlign.Left}
+              >
+                <ul style={{ listStyle: 'disc' }}>
+                  <li>
+                    <Text variant={TextVariant.bodyMd}>
+                      {t('hardwareWalletSubmissionWarningStep1', [
+                        hardwareWalletName,
+                      ])}
+                    </Text>
+                  </li>
+                  <li>
+                    <Text variant={TextVariant.bodyMd}>
+                      {t('hardwareWalletSubmissionWarningStep2', [
+                        hardwareWalletName,
+                      ])}
+                    </Text>
+                  </li>
+                </ul>
+              </BannerAlert>
+            )}
           {isNoQuotesAvailable && !isQuoteExpired && (
             <BannerAlert
               marginInline={4}
@@ -817,7 +862,7 @@ const PrepareBridgePage = () => {
           {isCannotVerifyTokenBannerOpen &&
             isEvm &&
             toToken &&
-            toTokenIsNotDefault &&
+            toTokenIsNotNative &&
             occurrences < 2 && (
               <BannerAlert
                 severity={BannerAlertSeverity.Warning}
@@ -829,7 +874,7 @@ const PrepareBridgePage = () => {
                 onClose={() => setIsCannotVerifyTokenBannerOpen(false)}
               />
             )}
-          {isEstimatedReturnLow && isLowReturnBannerOpen && (
+          {isEstimatedReturnLow && isLowReturnBannerOpen && activeQuote && (
             <BannerAlert
               ref={insufficientBalanceBannerRef}
               marginInline={4}
@@ -887,26 +932,6 @@ const PrepareBridgePage = () => {
               ])}
               textAlign={TextAlign.Left}
               onClose={() => setIsLowReturnBannerOpen(false)}
-            />
-          )}
-          {tokenAlert && isTokenAlertBannerOpen && (
-            <BannerAlert
-              ref={tokenAlertBannerRef}
-              marginInline={4}
-              marginBottom={3}
-              title={tokenAlert.titleId ? t(tokenAlert.titleId) : ''}
-              severity={
-                tokenAlert.type === TokenFeatureType.MALICIOUS
-                  ? BannerAlertSeverity.Danger
-                  : BannerAlertSeverity.Warning
-              }
-              description={
-                tokenAlert.descriptionId
-                  ? t(tokenAlert.descriptionId)
-                  : tokenAlert.description
-              }
-              textAlign={TextAlign.Left}
-              onClose={() => setIsTokenAlertBannerOpen(false)}
             />
           )}
           {!isLoading &&